apps/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20
httpx==0.28.1
python-jose[cryptography]==3.3.0
supabase==2.11.0
groq==0.13.1

# Jobs
celery==5.4.0
redis==5.2.1

